Move getAutoMatrixFor into own utils file.
